+++ Changelog.md
@@ -38,6 +38,7 @@
 - Fix: Players are able to mount Security Cameras that have been shut down by an EMP from another mod
 - Fix: A Portable Radar configured to not send repeating messages still repeats messages when multiple players are in its range
 - Fix: Mine remote access tools automatically remove positions of mines that are no longer in the world from their list
+- Fix: Some reinforced blocks can conduct redstone while their vanilla counterpart cannot do so
 
 --------------------------Changelog for v1.9.9 of SecurityCraft--------------------------